Fix how comment and like counts are displayed on the My Photos page
Correctly cast SQL counts to integers on the server and use Number() coercion in frontend reduce operations to ensure accurate display of like and comment totals.

+++ artifacts/api-server/src/routes/photos.ts
@@ -45,7 +45,7 @@ function buildPhotoSelect() {
       themeId: photosTable.themeId,
       themeName: themesTable.name,
       likeCount: photosTable.likeCount,
-      commentCount: sql<number>`coalesce(${commentCountSq.commentCount}, 0)`,
+      commentCount: sql<number>`coalesce(${commentCountSq.commentCount}, 0)::int`,
       createdAt: photosTable.createdAt,
     })
     .from(photosTable)
